refactor daemon scheduling test ownership
